refactor: customer and admin menus now use the command design pattern
